* 'master.newKite' of git.sj.koding.com:koding/koding:
  container: implement run test. fix container.run and polish code.
  container: distinguis between prepare and unprepare to be used with -run flag
  container: finish all prepare and unprepare tests. refactor unprepare.
  container: initialize test variables just once
  container: add test for adding and deleting static route
  container: add ebtables adding and removing test together with CheckEbtables method.
  container: tests for mount and umounts are finished.  Pure awesomeness: http://d.pr/i/wOKJ
  container: add mounting tests together with CheckMount method
  container: refactor creating container and overlaydir logic into functions for easy testing
  container: add a new test for webdir
  container: implement run and start commands
  container: remove unused code, change function names
  supervisor: improvements for lxc 1.0.0.alpha1

+++ go/src/koding/kites/supervisor/container/prepare.go
@@ -5,7 +5,7 @@ import (
 	"koding/kites/supervisor/rbd"
 	"os"
 	"os/exec"
-	"path"
+	"regexp"
 	"strconv"
 	"syscall"
 	"time"
@@ -15,66 +15,49 @@ import (
 // with lxc.start. We don't use lxc.create (which uses shell scipts for
 // templating). Instead we use this method which basically let us do things
 // more simpler. It creates the home directory, generates files like lxc.conf
-// and mounts the necessary disks.
+// and mounts the necessary filesystems.
 func (c *Container) Prepare() error {
-	fmt.Println("generating lxc files")
-	c.AsHost().PrepareDir(c.Path(""))
-	c.AsHost().GenerateFile(c.Path("config"), "config")
-	c.AsHost().GenerateFile(c.Path("fstab"), "fstab")
-	c.AsHost().GenerateFile(c.Path("ip-address"), "ip-address")
-
-	fmt.Println("mounting rbd")
-	err := c.MountRBD()
+	err := c.CreateContainerDir()
 	if err != nil {
 		return err
 	}
 
-	fmt.Println("generating  overlay files")
-	c.AsContainer().PrepareDir(c.OverlayPath(""))            //for chown
-	c.AsContainer().PrepareDir(c.OverlayPath("/lost+found")) // for chown
-	c.AsContainer().PrepareDir(c.OverlayPath("/etc"))
-
-	c.AsContainer().GenerateFile(c.OverlayPath("/etc/hostname"), "hostname")
-	c.AsContainer().GenerateFile(c.OverlayPath("/etc/hosts"), "hosts")
-	c.AsContainer().GenerateFile(c.OverlayPath("/etc/ldap.conf"), "ldap.conf")
-
-	fmt.Println("merging files")
-	c.MergePasswdFile()
-	c.MergeGroupFile()
-	c.MergeDpkgDatabase()
-
-	fmt.Println("preparing rootfs")
-	err = c.AsContainer().PrepareDir(c.Path("rootfs"))
+	err = c.MountRBD()
 	if err != nil {
 		return err
 	}
 
-	fmt.Println("mounting aufs")
+	err = c.CreateOverlay()
+	if err != nil {
+		return err
+	}
+
+	err = c.MergeFiles()
+	if err != nil {
+		return err
+	}
+
 	err = c.MountAufs()
 	if err != nil {
 		return err
 	}
 
-	fmt.Println("mounting pts")
-	err = c.MountPts()
+	err = c.PrepareAndMountPts()
 	if err != nil {
 		return err
 	}
 
-	fmt.Println("add ebtables")
 	err = c.AddEbtablesRule()
 	if err != nil {
 		return err
 	}
 
-	fmt.Println("add static rout")
 	err = c.AddStaticRoute()
 	if err != nil {
 		return err
 	}
 
-	fmt.Println("prepare home directories")
-	err = c.PrepareHomeDirectories()
+	err = c.PrepareHomeDirectory()
 	if err != nil {
 		return err
 	}
@@ -82,10 +65,22 @@ func (c *Container) Prepare() error {
 	return nil
 }
 
+func (c *Container) CreateContainerDir() error {
+	err := c.AsHost().PrepareDir(c.Path(""))
+	if err != nil {
+		return err // for now just check this, no need for others
+	}
+
+	c.AsHost().GenerateFile(c.Path("config"), "config")
+	c.AsHost().GenerateFile(c.Path("fstab"), "fstab")
+	c.AsHost().GenerateFile(c.Path("ip-address"), "ip-address")
+
+	return nil
+}
+
 func (c *Container) MountRBD() error {
 	r := rbd.NewRBD(c.Name)
 
-	fmt.Println("getting info")
 	out, err := r.Info(c.Name)
 	if err != nil {
 		return err
@@ -94,7 +89,6 @@ func (c *Container) MountRBD() error {
 	makeFileSystem := false
 	// means image doesn't exist, create new one
 	if out == nil {
-		fmt.Println("creating new one")
 		out, err := r.Create(c.Name, "1024")
 		if err != nil {
 			return fmt.Errorf("mountrbd create failed.", err, out)
@@ -103,14 +97,11 @@ func (c *Container) MountRBD() error {
 		makeFileSystem = true
 	}
 
-	fmt.Println("mapping rbd")
 	out, err = r.Map(c.Name)
 	if err != nil {
 		return fmt.Errorf("mountrbd map failed.", err, out)
 	}
-	fmt.Println("mapping out", string(out))
 
-	fmt.Println("waiting for", r.Device)
 	timeout := time.Now().Add(30 * time.Second)
 	for {
 		_, err := os.Stat(r.Device)
@@ -148,6 +139,15 @@ func (c *Container) MountRBD() error {
 		return err
 	}
 
+	mounted, err := c.CheckMount(mountDir)
+	if err != nil {
+		return err
+	}
+
+	if mounted {
+		return nil
+	}
+
 	if out, err := exec.Command("/bin/mount", "-t", "ext4", r.Device, mountDir).CombinedOutput(); err != nil {
 		os.Remove(mountDir)
 		return fmt.Errorf("mount rbd failed. err: %s\nout:%s\n", err, string(out))
@@ -156,11 +156,83 @@ func (c *Container) MountRBD() error {
 	return nil
 }
 
-func (c *Container) MountAufs() error {
-	// if out, err := exec.Command("/bin/mount", "--no-mtab", "-t", "overlayfs", "-o", fmt.Sprintf("lowerdir=%s,upperdir=%s", vm.LowerdirFile("/"), vm.OverlayFile("/")), "overlayfs", vm.File("rootfs")).CombinedOutput(); err != nil {
+// CheckMount returns true if the given path/filesystem is mounted.
+func (c *Container) CheckMount(path string) (bool, error) {
+	// mountpoint returns with exit status 0 when the mountpoint is available
+	// or it returns exit status 1.
+	cmd := exec.Command("/bin/mountpoint", "-q", path+"/")
+	err := cmd.Run()
+	if err == nil {
+		return true, nil // exit status 0, means mounted
+	}
 
-	// mount "/var/lib/lxc/vm-{id}/overlay" (rw) and "/var/lib/lxc/vmroot" (ro)
-	// under "/var/lib/lxc/vm-{id}/rootfs"
+	_, ok := err.(*exec.ExitError)
+	if !ok {
+		// means mountpoint couldn't be invoked or the binary is not
+		// avilable. Check what the returning err is saying.
+		return false, err
+	}
+
+	return false, nil // exit status 1, means not mounted
+}
+
+func (c *Container) CheckExt4(device string) error {
+	// check/correct filesystem
+	out, err := exec.Command("/sbin/fsck.ext4", "-p", device).CombinedOutput()
+	if err == nil {
+		return nil
+	}
+
+	exitError, ok := err.(*exec.ExitError)
+	if ok || exitError.Sys().(syscall.WaitStatus).ExitStatus() != 4 {
+		return fmt.Errorf("fsck.ext4 failed %s. %s %s", c.HostnameAlias, err, string(out))
+	}
+
+	out, err = exec.Command("/sbin/fsck.ext4", "-y", device).CombinedOutput()
+	if err == nil {
+		return nil
+	}
+
+	exitError, ok = err.(*exec.ExitError)
+	if !ok || exitError.Sys().(syscall.WaitStatus).ExitStatus() != 1 {
+		return fmt.Errorf("fsck.ext4 could not automatically repair FS for %s. err '%s', out: '%s;",
+			c.HostnameAlias, err, string(out))
+	}
+
+	return nil
+}
+
+func (c *Container) CreateOverlay() error {
+	err := c.AsContainer().PrepareDir(c.OverlayPath("")) //for chown
+	if err != nil {
+		return err // for now just check this, no need for others
+	}
+
+	c.AsContainer().PrepareDir(c.OverlayPath("lost+found")) // for chown
+	c.AsContainer().PrepareDir(c.OverlayPath("etc"))
+
+	c.AsContainer().GenerateFile(c.OverlayPath("etc/hostname"), "hostname")
+	c.AsContainer().GenerateFile(c.OverlayPath("etc/hosts"), "hosts")
+	c.AsContainer().GenerateFile(c.OverlayPath("etc/ldap.conf"), "ldap.conf")
+
+	return nil
+}
+
+func (c *Container) MergeFiles() error {
+	c.MergePasswdFile()
+	c.MergeGroupFile()
+	c.MergeDpkgDatabase()
+	return nil
+}
+
+// mount "/var/lib/lxc/vm-{id}/overlay" (rw) and "/var/lib/lxc/vmroot" (ro)
+// under "/var/lib/lxc/vm-{id}/rootfs"
+func (c *Container) MountAufs() error {
+	err := c.AsContainer().PrepareDir(c.Path("rootfs"))
+	if err != nil {
+		return err
+	}
+
 	out, err := exec.Command("/bin/mount", "--no-mtab", "-t", "aufs", "-o",
 		fmt.Sprintf("noplink,br=%s:%s", c.OverlayPath(""), vmRoot+"/rootfs/"),
 		"aufs", c.Path("rootfs")).CombinedOutput()
@@ -171,8 +243,11 @@ func (c *Container) MountAufs() error {
 	return nil
 }
 
-func (c *Container) MountPts() error {
-	c.AsContainer().PrepareDir(c.PtsDir())
+func (c *Container) PrepareAndMountPts() error {
+	err := c.AsContainer().PrepareDir(c.PtsDir())
+	if err != nil {
+		return err
+	}
 
 	out, err := exec.Command("/bin/mount", "--no-mtab", "-t", "devpts", "-o",
 		"rw,noexec,nosuid,newinstance,gid="+strconv.Itoa(RootUIDOffset+5)+",mode=0620,ptmxmode=0666",
@@ -187,8 +262,17 @@ func (c *Container) MountPts() error {
 	return nil
 }
 
+// AddEbtablesRule adds entries to restrict IP and MAC
 func (c *Container) AddEbtablesRule() error {
-	// add ebtables entry to restrict IP and MAC
+	available, err := c.CheckEbtables()
+	if err != nil {
+		return err
+	}
+
+	if available {
+		return nil
+	}
+
 	out, err := exec.Command("/sbin/ebtables", "--append", "VMS", "--protocol", "IPv4", "--source",
 		c.MAC().String(), "--ip-src", c.IP.String(), "--in-interface", c.VEth(),
 		"--jump", "ACCEPT").CombinedOutput()
@@ -199,8 +283,17 @@ func (c *Container) AddEbtablesRule() error {
 	return nil
 }
 
+//AddStaticRoute adds a route so it is redistributed by BGP
 func (c *Container) AddStaticRoute() error {
-	// add a static route so it is redistributed by BGP
+	available, err := c.CheckStaticRoute()
+	if err != nil {
+		return err
+	}
+
+	if available {
+		return nil
+	}
+
 	out, err := exec.Command("/sbin/route", "add", c.IP.String(), "lxcbr0").CombinedOutput()
 	if err != nil {
 		return fmt.Errorf("adding route failed. err: %s\n out:%s\n", err, out)
@@ -209,7 +302,17 @@ func (c *Container) AddStaticRoute() error {
 	return nil
 }
 
-func (c *Container) PrepareHomeDirectories() error {
+func (c *Container) CheckStaticRoute() (bool, error) {
+	cmd := exec.Command("/sbin/ip", "route", "show", c.IP.String())
+	out, err := cmd.CombinedOutput()
+	if err != nil {
+		return false, err
+	}
+
+	return regexp.Match(c.IP.String(), out)
+}
+
+func (c *Container) PrepareHomeDirectory() error {
 	// make sure that executable flag is set
 	os.Chmod(c.Path("rootfs/"), 0755)
 	os.Chmod(c.Path("rootfs/home"), 0755)
@@ -232,21 +335,17 @@ func (c *Container) PrepareHomeDirectories() error {
 func (c *Container) createUserHome() error {
 	homeDir := c.Path(fmt.Sprintf("rootfs/home/%s", c.Username))
 
-	fmt.Println("creating user home:...", homeDir)
 	if info, err := os.Stat(homeDir); !os.IsNotExist(err) {
 		// make sure that user read and executable flag is set
-		fmt.Println("username exist", info.Name(), info.IsDir())
 		return os.Chmod(homeDir, info.Mode().Perm()|0511)
 	}
 
-	fmt.Println("home directory does not exist", c.Username)
 	// home directory does not exist, create it
 	err := os.MkdirAll(homeDir, 0755)
 	if err != nil {
 		return err
 	}
 
-	fmt.Println("chowning ", homeDir)
 	err = c.AsUser().Lchown(homeDir)
 	if err != nil {
 		return err
@@ -270,7 +369,7 @@ func (c *Container) createUserHome() error {
 // Apache webserver.
 func (c *Container) createWebDir(webDir string) error {
 	wwwDir := c.Path("rootfs/var/www")
-	if err := os.Symlink(webDir, "/var/www"); err != nil && !os.IsExist(err) {
+	if err := os.Symlink(webDir, wwwDir); err != nil && !os.IsExist(err) {
 		return err
 	}
 	c.AsContainer().Lchown(wwwDir)
@@ -313,7 +412,6 @@ func (c *Container) copyIntoContainer(src, dst string) error {
 	if fi.Name() == "empty-directory" {
 		// ignored file
 	} else if fi.IsDir() {
-		fmt.Println("src is a dir:", src, path.Base(src))
 		err := os.Mkdir(dst, fi.Mode())
 		if err != nil && !os.IsExist(err) {
 			return err
@@ -335,36 +433,10 @@ func (c *Container) copyIntoContainer(src, dst string) error {
 			}
 		}
 	} else {
-		fmt.Printf("copying from '%s' to '%s'", src, dst)
 		err := c.AsUser().CopyFile(src, dst)
 		if err != nil {
 			return err
 		}
-	}
-
-	return nil
-}
-
-func (c *Container) CheckExt4(device string) error {
-	// check/correct filesystem
-	out, err := exec.Command("/sbin/fsck.ext4", "-p", device).CombinedOutput()
-	if err == nil {
-		return nil
-	}
-
-	exitError, ok := err.(*exec.ExitError)
-	if !ok || exitError.Sys().(syscall.WaitStatus).ExitStatus() == 4 {
-		return fmt.Errorf(fmt.Sprintf("fsck.ext4 failed %s. %s %s", c.HostnameAlias), err, string(out))
-	}
-
-	out, err = exec.Command("/sbin/fsck.ext4", "-y", device).CombinedOutput()
-	if err == nil {
-		return nil
-	}
-
-	exitError, ok = err.(*exec.ExitError)
-	if !ok || exitError.Sys().(syscall.WaitStatus).ExitStatus() != 1 {
-		return fmt.Errorf(fmt.Sprintf("fsck.ext4 could not automatically repair FS for %s.", c.HostnameAlias), err, string(out))
 	}
 
 	return nil
